Updated total shares logic
Broke out the total shares logic into it's own protected function to
determine if, when, and where the total shares should display. Changed
the open and close wrapper adder to a single add_html_wrapper function
that does both at the same time and takes care of the floating wrapper
as well reducing the code needed for the floating buttons.